Correct another instance of bf758bf
Actually by mistake the previous commit doesn't correct the line of code
I was having trouble with, it was in a different section which was not
run. In any case, the line of code is identical so I assume that other
instance would have thrown an error had it run. All instances (at least
in the extract_and_filter files) should be fixed now.