Add function for overlaying contour plot on spots
This function is identical to `iss_change_plot()` except that it adds
a semi-transparent contour plot that is overlayed atop the spots to show
their density. Note that the color map is not absolute, so density
intensities are not directly comparable between plots, although relative
spatial distribution should be be. This can be fixed in a future commit.